requirements.txt
astroid==2.2.5
asyncio-nats-client==0.9.2
isort==4.3.21
lazy-object-proxy==1.4.2
mccabe==0.6.1
pylint==2.3.1
six==1.12.0
typed-ast==1.4.0
wrapt==1.11.2
